[refactor] #25 - 그룹 상세 화면 뷰 파일 분리
